Move slider classification which is also flaky
