+++ skyportal/tests/flaky/test_frontend_flaky.py
@@ -4,6 +4,8 @@ import uuid
 import pandas as pd
 import time
 from regions import Regions
+from tdtax import taxonomy, __version__
+
 import pytest
 from skyportal.tests import api
 from selenium.webdriver.common.keys import Keys
@@ -986,3 +988,51 @@ def test_user_expiration(
     driver.get(f'/become_user/{user.id}')
     driver.get("/")
     driver.wait_for_xpath_to_disappear("//*[contains(text(), 'Top Sources')]")
+
+
+def test_slider_classifications(
+    driver,
+    public_source,
+    super_admin_user,
+    public_group,
+    taxonomy_token,
+):
+    test_taxonomy = "test taxonomy" + str(uuid.uuid4())
+    status, _ = api(
+        'POST',
+        'taxonomy',
+        data={
+            'name': test_taxonomy,
+            'hierarchy': taxonomy,
+            'group_ids': [public_group.id],
+            'provenance': f"tdtax_{__version__}",
+            'version': __version__,
+            'isLatest': True,
+        },
+        token=taxonomy_token,
+    )
+    assert status == 200
+
+    driver.get(f"/become_user/{super_admin_user.id}")  # become a super-user
+
+    # Go to the group sources page
+    driver.get(f"/group_sources/{public_group.id}")
+
+    # Wait for the group name appears
+    driver.wait_for_xpath(f"//*[text()[contains(., '{public_group.name}')]]")
+
+    # Expand public source row
+    driver.click_xpath("//*[@id='expandable-button']")
+
+    # Select test taxonomy
+    driver.click_xpath(f"//*[@id='taxonomy-select-{public_source.id}']")
+    driver.click_xpath(f"//li[text()='{test_taxonomy}']", scroll_parent=True)
+    driver.click_xpath("//h6[text()='Stellar variable']")
+
+    # Set 'Algol' slider to 0.5
+    driver.click_xpath("//span[@id='Algol']//span[@data-index='2']")
+    driver.click_xpath("//button[@name='submitClassificationsButton']")
+
+    # need to reload the page to see classification
+    driver.get(f"/group_sources/{public_group.id}")
+    driver.wait_for_xpath(f"//*[text()[contains(., '{'Algol'}')]]")
